Updated Camera logic + work on Sound Detector
Spotting Camera now rotates correctly without hardcoding and detects if player is in it's vision.

Sound Detector is currently a proof of concept but any action that makes noise will send a message to the sound detector that then checks its distance and multiplies it by the severity of the noise. This results in a counter getting changed which if it gets too high will set off the alarm. Any noise too loud or too close will immediately set off the alarm.